requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
configparser==7.2.0
DateTime==5.5
defusedxml==0.7.1
et_xmlfile==2.0.0
idna==3.10
jira==3.10.5
logging==0.4.9.6
lxml==5.4.0
numpy==2.2.6
oauthlib==3.2.2
openpyxl==3.2.0b1
packaging==25.0
pandas==2.2.3
pbr==6.1.1
pillow==11.2.1
pygerrit2==2.0.15
python-dateutil==2.9.0
python-docx==1.1.2
pytz==2025.2
regex==2024.11.6
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
requests==2.32.3
setuptools==80.9.0
six==1.17.0
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
zope.interface==7.2
pytest==8.3.4